knew factors for the guild database

+++ src/commands/Moderation/purge.js
@@ -22,17 +22,27 @@ module.exports = {
 			.setTitle('The messages are way too old')
 			.setColor(0xff0000);
 
-		if (!message.member.hasPermission('MANAGE_GUILD'))
-			return noPerms(message, 'MANAGE_GUILD');
+		if (
+			!message.member.hasPermission(
+				'MANAGE_GUILD' ||
+				'MANAGE_CHANNELS' ||
+				'MANAGE_MESSAGES'
+			)
+		)
+			return noPerms(message, 'MANAGE_MESSAGES');
 
 		if (!args[0]) return message.channel.send(noargs);
 		message.channel
 			.bulkDelete(args[0])
 			.then(() => {
-				message.channel.send(deletedmessages).then((msg) => msg.delete(10000));
+				message.channel
+					.send(deletedmessages)
+					.then((msg) => msg.delete(10000));
 			})
 			.catch((error) => {
-				message.channel.send(oldmessage).then((msg) => msg.delete(10000));
+				message.channel
+					.send(oldmessage)
+					.then((msg) => msg.delete(10000));
 			});
 	}
 };
